Put workflow tests back in data_sources for now
