+++ x-pack/platform/plugins/shared/agent_builder/tsconfig.json
@@ -117,9 +117,5 @@
     "@kbn/fs",
     "@kbn/react-hooks",
     "@kbn/search-inference-endpoints",
-    "@kbn/connector-specs",
-    "@kbn/workflows-execution-engine",
-    "@kbn/tooling-log",
-    "@kbn/test",
   ]
 }
