+++ x-pack/platform/plugins/shared/data_sources/server/sources/google_calendar/workflows.test.ts
@@ -13,7 +13,7 @@ import {
   loadWorkflowsFromConnectorSpec,
   registerExtensionSteps,
   type ProcessedWorkflow,
-} from './__test_helpers__/workflow.test_helpers';
+} from '../workflow.test_helpers';
 
 const CONNECTOR_NAME = 'fake-google-calendar-connector';
 const CONNECTOR_ID = 'fake-gcal-connector-uuid';
