+++ x-pack/platform/plugins/shared/data_sources/server/sources/salesforce/workflows.test.ts
@@ -12,17 +12,17 @@ import {
   getWorkflowYaml,
   loadWorkflowsFromConnectorSpec,
   type ProcessedWorkflow,
-} from './__test_helpers__/workflow.test_helpers';
+} from '../workflow.test_helpers';
 
-const CONNECTOR_NAME = 'fake-firecrawl-connector';
-const CONNECTOR_ID = 'fake-firecrawl-connector-uuid';
+const CONNECTOR_NAME = 'fake-salesforce-connector';
+const CONNECTOR_ID = 'fake-salesforce-connector-uuid';
 
-describe('firecrawl workflows', () => {
+describe('salesforce workflows', () => {
   let fixture: WorkflowRunFixture;
   let workflows: ProcessedWorkflow[];
 
   beforeAll(() => {
-    workflows = loadWorkflowsFromConnectorSpec('.firecrawl', {
+    workflows = loadWorkflowsFromConnectorSpec('.salesforce', {
       connectorName: CONNECTOR_NAME,
     });
   });
@@ -31,7 +31,7 @@ describe('firecrawl workflows', () => {
     fixture = new WorkflowRunFixture();
 
     fixture.scopedActionsClientMock.getAll.mockResolvedValue([
-      { id: CONNECTOR_ID, name: CONNECTOR_NAME, actionTypeId: '.firecrawl' },
+      { id: CONNECTOR_ID, name: CONNECTOR_NAME, actionTypeId: '.salesforce' },
     ]);
 
     fixture.scopedActionsClientMock.returnMockedConnectorResult = async ({
@@ -44,40 +44,44 @@ describe('firecrawl workflows', () => {
       const subAction = params.subAction as string;
 
       switch (subAction) {
-        case 'scrape':
-          return {
-            status: 'ok',
-            actionId,
-            data: { markdown: '# Example\nContent here', metadata: { title: 'Example' } },
-          };
         case 'search':
           return {
             status: 'ok',
             actionId,
-            data: [{ url: 'https://example.com', markdown: '# Result', metadata: {} }],
+            data: { searchRecords: [{ Id: '001', Name: 'Acme' }] },
           };
-        case 'map':
+        case 'query':
           return {
             status: 'ok',
             actionId,
-            data: { links: ['https://example.com/p1', 'https://example.com/p2'] },
+            data: { records: [{ Id: '001', Name: 'Acme' }], totalSize: 1, done: true },
           };
-        case 'crawl':
-          return { status: 'ok', actionId, data: { id: 'crawl-job-123', status: 'scraping' } };
-        case 'getCrawlStatus':
+        case 'list_records':
           return {
             status: 'ok',
             actionId,
-            data: { status: 'completed', total: 5, completed: 5, data: [] },
+            data: { records: [{ Id: '001' }], totalSize: 1, done: true },
           };
-        case 'crawlAndWait':
+        case 'get_record':
           return {
             status: 'ok',
             actionId,
-            data: { status: 'completed', total: 3, completed: 3, data: [] },
+            data: { Id: '001', Name: 'Acme Corp', Type: 'Customer' },
+          };
+        case 'download_file':
+          return {
+            status: 'ok',
+            actionId,
+            data: { base64: 'dGVzdA==', contentType: 'application/pdf' },
+          };
+        case 'describe':
+          return {
+            status: 'ok',
+            actionId,
+            data: { name: 'Account', fields: [{ name: 'Id', type: 'id' }] },
           };
         default:
-          throw new Error(`Unexpected Firecrawl subAction: ${subAction}`);
+          throw new Error(`Unexpected Salesforce subAction: ${subAction}`);
       }
     };
   });
@@ -94,36 +98,11 @@ describe('firecrawl workflows', () => {
     }
   });
 
-  describe('scrape workflow', () => {
-    it('forwards scrape parameters to the connector', async () => {
-      await fixture.runWorkflow({
-        workflowYaml: getWorkflowYaml(workflows, 'scrape'),
-        inputs: { url: 'https://example.com' },
-      });
-
-      expect(getWorkflowExecution()?.status).toBe(ExecutionStatus.COMPLETED);
-
-      expect(fixture.scopedActionsClientMock.execute).toHaveBeenCalledWith(
-        expect.objectContaining({
-          params: expect.objectContaining({
-            subAction: 'scrape',
-            subActionParams: {
-              url: 'https://example.com',
-              onlyMainContent: true,
-              waitFor: 0,
-              maxMarkdownLength: 100000,
-            },
-          }),
-        })
-      );
-    });
-  });
-
   describe('search workflow', () => {
     it('forwards search parameters to the connector', async () => {
       await fixture.runWorkflow({
         workflowYaml: getWorkflowYaml(workflows, 'search'),
-        inputs: { query: 'kibana plugins', limit: 3 },
+        inputs: { search_term: 'Acme Corp', returning: 'Account,Contact' },
       });
 
       expect(getWorkflowExecution()?.status).toBe(ExecutionStatus.COMPLETED);
@@ -133,8 +112,9 @@ describe('firecrawl workflows', () => {
           params: expect.objectContaining({
             subAction: 'search',
             subActionParams: {
-              query: 'kibana plugins',
-              limit: 3,
+              searchTerm: 'Acme Corp',
+              returning: 'Account,Contact',
+              nextRecordsUrl: undefined,
             },
           }),
         })
@@ -142,11 +122,11 @@ describe('firecrawl workflows', () => {
     });
   });
 
-  describe('map workflow', () => {
-    it('forwards map parameters to the connector', async () => {
+  describe('query workflow', () => {
+    it('forwards SOQL query to the connector', async () => {
       await fixture.runWorkflow({
-        workflowYaml: getWorkflowYaml(workflows, 'map'),
-        inputs: { url: 'https://example.com', search: 'docs' },
+        workflowYaml: getWorkflowYaml(workflows, 'query'),
+        inputs: { soql: 'SELECT Id, Name FROM Account LIMIT 10' },
       });
 
       expect(getWorkflowExecution()?.status).toBe(ExecutionStatus.COMPLETED);
@@ -154,12 +134,10 @@ describe('firecrawl workflows', () => {
       expect(fixture.scopedActionsClientMock.execute).toHaveBeenCalledWith(
         expect.objectContaining({
           params: expect.objectContaining({
-            subAction: 'map',
+            subAction: 'query',
             subActionParams: {
-              url: 'https://example.com',
-              search: 'docs',
-              limit: 5000,
-              includeSubdomains: true,
+              soql: 'SELECT Id, Name FROM Account LIMIT 10',
+              nextRecordsUrl: undefined,
             },
           }),
         })
@@ -167,11 +145,11 @@ describe('firecrawl workflows', () => {
     });
   });
 
-  describe('crawl workflow', () => {
-    it('crawl action starts an async crawl job', async () => {
+  describe('list_records workflow', () => {
+    it('forwards list parameters to the connector', async () => {
       await fixture.runWorkflow({
-        workflowYaml: getWorkflowYaml(workflows, 'crawl'),
-        inputs: { crawl_action: 'crawl', url: 'https://example.com', limit: 10 },
+        workflowYaml: getWorkflowYaml(workflows, 'list_records'),
+        inputs: { sobject_name: 'Account', limit: 5 },
       });
 
       expect(getWorkflowExecution()?.status).toBe(ExecutionStatus.COMPLETED);
@@ -179,22 +157,23 @@ describe('firecrawl workflows', () => {
       expect(fixture.scopedActionsClientMock.execute).toHaveBeenCalledWith(
         expect.objectContaining({
           params: expect.objectContaining({
-            subAction: 'crawl',
+            subAction: 'list_records',
             subActionParams: {
-              url: 'https://example.com',
-              limit: 10,
-              maxDiscoveryDepth: undefined,
-              allowExternalLinks: false,
+              sobjectName: 'Account',
+              limit: 5,
+              nextRecordsUrl: undefined,
             },
           }),
         })
       );
     });
+  });
 
-    it('getCrawlStatus checks an existing crawl job', async () => {
+  describe('get_record workflow', () => {
+    it('forwards record lookup to the connector', async () => {
       await fixture.runWorkflow({
-        workflowYaml: getWorkflowYaml(workflows, 'crawl'),
-        inputs: { crawl_action: 'getCrawlStatus', id: 'crawl-job-123' },
+        workflowYaml: getWorkflowYaml(workflows, 'get_record'),
+        inputs: { sobject_name: 'Account', record_id: '001ABC123' },
       });
 
       expect(getWorkflowExecution()?.status).toBe(ExecutionStatus.COMPLETED);
@@ -202,19 +181,22 @@ describe('firecrawl workflows', () => {
       expect(fixture.scopedActionsClientMock.execute).toHaveBeenCalledWith(
         expect.objectContaining({
           params: expect.objectContaining({
-            subAction: 'getCrawlStatus',
+            subAction: 'get_record',
             subActionParams: {
-              id: 'crawl-job-123',
+              sobjectName: 'Account',
+              recordId: '001ABC123',
             },
           }),
         })
       );
     });
+  });
 
-    it('crawlAndWait runs a synchronous crawl', async () => {
+  describe('download_file workflow', () => {
+    it('forwards content version ID to the connector', async () => {
       await fixture.runWorkflow({
-        workflowYaml: getWorkflowYaml(workflows, 'crawl'),
-        inputs: { crawl_action: 'crawlAndWait', url: 'https://example.com' },
+        workflowYaml: getWorkflowYaml(workflows, 'download_file'),
+        inputs: { content_version_id: '068ABC123' },
       });
 
       expect(getWorkflowExecution()?.status).toBe(ExecutionStatus.COMPLETED);
@@ -222,12 +204,31 @@ describe('firecrawl workflows', () => {
       expect(fixture.scopedActionsClientMock.execute).toHaveBeenCalledWith(
         expect.objectContaining({
           params: expect.objectContaining({
-            subAction: 'crawlAndWait',
+            subAction: 'download_file',
             subActionParams: {
-              url: 'https://example.com',
-              limit: 20,
-              maxDiscoveryDepth: undefined,
-              allowExternalLinks: false,
+              contentVersionId: '068ABC123',
+            },
+          }),
+        })
+      );
+    });
+  });
+
+  describe('describe workflow', () => {
+    it('forwards sobject name to the connector', async () => {
+      await fixture.runWorkflow({
+        workflowYaml: getWorkflowYaml(workflows, 'describe'),
+        inputs: { sobject_name: 'Account' },
+      });
+
+      expect(getWorkflowExecution()?.status).toBe(ExecutionStatus.COMPLETED);
+
+      expect(fixture.scopedActionsClientMock.execute).toHaveBeenCalledWith(
+        expect.objectContaining({
+          params: expect.objectContaining({
+            subAction: 'describe',
+            subActionParams: {
+              sobjectName: 'Account',
             },
           }),
         })
