+++ x-pack/platform/plugins/shared/data_sources/server/sources/sharepoint_online/workflows.test.ts
@@ -12,7 +12,7 @@ import {
   getWorkflowYaml,
   loadWorkflowsFromConnectorSpec,
   type ProcessedWorkflow,
-} from './__test_helpers__/workflow.test_helpers';
+} from '../workflow.test_helpers';
 
 const CONNECTOR_NAME = 'fake-sharepoint-connector';
 const CONNECTOR_ID = 'fake-sp-connector-uuid';
