+++ x-pack/platform/plugins/shared/data_sources/server/sources/zoom/workflows.test.ts
@@ -12,7 +12,7 @@ import {
   getWorkflowYaml,
   loadWorkflowsFromConnectorSpec,
   type ProcessedWorkflow,
-} from './__test_helpers__/workflow.test_helpers';
+} from '../workflow.test_helpers';
 
 const CONNECTOR_NAME = 'fake-zoom-connector';
 const CONNECTOR_ID = 'fake-zoom-connector-uuid';
