+++ x-pack/platform/plugins/shared/data_sources/tsconfig.json
@@ -45,6 +45,8 @@
     "@kbn/task-manager-plugin",
     "@kbn/workflows",
     "@kbn/workflows-execution-engine",
+    "@kbn/tooling-log",
+    "@kbn/test",
     "@kbn/workflows-extensions"
   ]
 }
